+++ src/api/todos.ts
@@ -1,3 +1,4 @@
+import { Filter } from '../types/Filter';
 import { Todo } from '../types/Todo';
 import { client } from '../utils/fetchClient';
 
@@ -5,6 +6,19 @@ export const USER_ID = 962;
 
 export const getTodos = () => {
   return client.get<Todo[]>(`/todos?userId=${USER_ID}`);
+};
+
+export const getFilteredTodos = (filter: Filter) => {
+  switch (filter) {
+    case 'all':
+      return client.get<Todo[]>(`/todos?userId=${USER_ID}`);
+    case 'active':
+      return client.get<Todo[]>(`/todos?userId=${USER_ID}&completed=false`);
+    case 'completed':
+      return client.get<Todo[]>(`/todos?userId=${USER_ID}&completed=true`);
+    default:
+      return client.get<Todo[]>(`/todos?userId=${USER_ID}`);
+  }
 };
 
 export const deleteTodo = (todoId: number) => {
